ssu2: split message types between handshake and data
