+++ emissary-core/src/transport/ssu2/message/mod.rs
@@ -21,32 +21,29 @@
 //! https://geti2p.net/spec/ssu2#noise-payload
 
 use crate::{
-    crypto::{
-        chachapoly::{ChaCha, ChaChaPoly},
-        EphemeralPublicKey, StaticPublicKey,
-    },
+    crypto::{chachapoly::ChaCha, EphemeralPublicKey},
     error::Ssu2Error,
     i2np::{Message, MessageType as I2npMessageType},
     primitives::{MessageId, RouterInfo},
-    runtime::Runtime,
-    transport::{ssu2::session::KeyContext, TerminationReason},
 };
 
-use bytes::{BufMut, Bytes, BytesMut};
+use bytes::{BufMut, BytesMut};
 use nom::{
     bytes::complete::take,
     error::{make_error, ErrorKind},
     number::complete::{be_u16, be_u32, be_u64, be_u8},
     Err, IResult,
 };
-use rand_core::RngCore;
 
-use alloc::{vec, vec::Vec};
+use alloc::vec::Vec;
 use core::{
     fmt,
     net::{IpAddr, SocketAddr},
     ops::{Deref, Range},
 };
+
+pub mod data;
+pub mod handshake;
 
 /// Logging target for the file.
 const LOG_TARGET: &str = "emissary::ssu2::message";
@@ -928,673 +925,6 @@ impl TryFrom<u8> for MessageType {
     }
 }
 
-/// Message kind for [`DataMessageBuilder`].
-enum MessageKind<'a> {
-    UnFragmented {
-        /// Unfragmented I2NP message.
-        message: &'a [u8],
-    },
-
-    /// First fragment.
-    FirstFragment {
-        /// Fragment.
-        fragment: &'a [u8],
-
-        /// Short expiration.
-        expiration: u32,
-
-        /// Message type.
-        message_type: I2npMessageType,
-
-        /// Message ID.
-        message_id: u32,
-    },
-
-    /// Follow-on fragment.
-    FollowOnFragment {
-        /// Fragment.
-        fragment: &'a [u8],
-
-        /// Fragment number.
-        fragment_num: u8,
-
-        /// Last fragment.
-        last: bool,
-
-        /// Message ID.
-        message_id: u32,
-    },
-}
-
-/// Data message
-#[derive(Default)]
-pub struct DataMessageBuilder<'a> {
-    /// ACK information.
-    acks: Option<(u32, u8, Option<Vec<(u8, u8)>>)>,
-
-    // Destination connection ID.
-    dst_id: Option<u64>,
-
-    /// Message kind.
-    i2np: Option<MessageKind<'a>>,
-
-    /// Key context for the message.
-    key_context: Option<([u8; 32], &'a KeyContext)>,
-
-    /// Payload length.
-    payload_len: usize,
-
-    /// Packet number.
-    pkt_num: Option<u32>,
-
-    /// Termination reason.
-    termination_reason: Option<TerminationReason>,
-}
-
-impl<'a> DataMessageBuilder<'a> {
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, value: u64) -> Self {
-        self.dst_id = Some(value);
-        self
-    }
-
-    /// Specify packet number.
-    pub fn with_pkt_num(mut self, value: u32) -> Self {
-        self.pkt_num = Some(value);
-        self
-    }
-
-    /// Specify key context.
-    pub fn with_key_context(mut self, intro_key: [u8; 32], key_ctx: &'a KeyContext) -> Self {
-        self.key_context = Some((intro_key, key_ctx));
-        self
-    }
-
-    /// Specify I2NP message.
-    pub fn with_i2np(mut self, message: &'a [u8]) -> Self {
-        self.payload_len = self
-            .payload_len
-            .saturating_add(1usize) // type
-            .saturating_add(2usize) // len
-            .saturating_add(message.len());
-        self.i2np = Some(MessageKind::UnFragmented { message });
-        self
-    }
-
-    /// Specify first fragment.
-    pub fn with_first_fragment(
-        mut self,
-        message_type: I2npMessageType,
-        message_id: u32,
-        expiration: u32,
-        fragment: &'a [u8],
-    ) -> Self {
-        self.payload_len = self
-            .payload_len
-            .saturating_add(1usize) // type
-            .saturating_add(2usize) // len
-            .saturating_add(fragment.len());
-        self.i2np = Some(MessageKind::FirstFragment {
-            expiration,
-            fragment,
-            message_id,
-            message_type,
-        });
-        self
-    }
-
-    /// Specify follow-on fragment.
-    pub fn with_follow_on_fragment(
-        mut self,
-        message_id: u32,
-        fragment_num: u8,
-        last: bool,
-        fragment: &'a [u8],
-    ) -> Self {
-        self.payload_len = self
-            .payload_len
-            .saturating_add(1usize) // type
-            .saturating_add(2usize) // len
-            .saturating_add(fragment.len());
-        self.i2np = Some(MessageKind::FollowOnFragment {
-            fragment,
-            fragment_num,
-            last,
-            message_id,
-        });
-        self
-    }
-
-    /// Specify ACK information.
-    pub fn with_ack(
-        mut self,
-        ack_through: u32,
-        num_acks: u8,
-        ranges: Option<Vec<(u8, u8)>>,
-    ) -> Self {
-        self.payload_len = self
-            .payload_len
-            .saturating_add(1usize) // type
-            .saturating_add(2usize) // len
-            .saturating_add(4usize) // ack through
-            .saturating_add(1usize) // num acks
-            .saturating_add(ranges.as_ref().map_or(0usize, |ranges| ranges.len() * 2)); // ranges
-        self.acks = Some((ack_through, num_acks, ranges));
-        self
-    }
-
-    /// Add termination block.
-    pub fn with_termination(mut self, termination_reason: TerminationReason) -> Self {
-        self.termination_reason = Some(termination_reason);
-        self
-    }
-
-    /// Build message into one or more packets.
-    pub fn build(mut self) -> BytesMut {
-        let pkt_num = self.pkt_num.expect("to exist");
-
-        let mut header = {
-            let mut out = BytesMut::with_capacity(16usize);
-
-            out.put_u64_le(self.dst_id.expect("to exist"));
-            out.put_u32(pkt_num);
-
-            out.put_u8(*MessageType::Data);
-            out.put_u8(0u8); // immediate ack
-            out.put_u16(0u16); // more flags
-
-            out
-        };
-
-        // build payload
-        let mut payload = {
-            let mut out = BytesMut::with_capacity(self.payload_len + POLY13055_MAC_LEN);
-
-            match self.i2np.take() {
-                None => {}
-                Some(MessageKind::UnFragmented { message }) => {
-                    out.put_u8(BlockType::I2Np.as_u8());
-                    out.put_slice(message);
-                }
-                Some(MessageKind::FirstFragment {
-                    expiration,
-                    fragment,
-                    message_id,
-                    message_type,
-                }) => {
-                    out.put_u8(BlockType::FirstFragment.as_u8());
-                    out.put_u16((fragment.len() + 1 + 4 + 4) as u16);
-                    out.put_u8(message_type.as_u8());
-                    out.put_u32(message_id);
-                    out.put_u32(expiration);
-                    out.put_slice(fragment);
-                }
-                Some(MessageKind::FollowOnFragment {
-                    fragment,
-                    fragment_num,
-                    last,
-                    message_id,
-                }) => {
-                    out.put_u8(BlockType::FollowOnFragment.as_u8());
-                    out.put_u16((fragment.len() + 1 + 4) as u16);
-                    out.put_u8((fragment_num << 1) | last as u8);
-                    out.put_u32(message_id);
-                    out.put_slice(fragment);
-                }
-            }
-
-            match self.acks.take() {
-                None => {}
-                Some((ack_through, num_acks, None)) => {
-                    out.put_u8(BlockType::Ack.as_u8());
-                    out.put_u16(5u16);
-                    out.put_u32(ack_through);
-                    out.put_u8(num_acks);
-                }
-                Some((ack_through, num_acks, Some(ranges))) => {
-                    out.put_u8(BlockType::Ack.as_u8());
-                    out.put_u16((5usize + ranges.len() * 2) as u16);
-                    out.put_u32(ack_through);
-                    out.put_u8(num_acks);
-
-                    ranges.into_iter().for_each(|(nack, ack)| {
-                        out.put_u8(nack);
-                        out.put_u8(ack);
-                    });
-                }
-            }
-
-            out.to_vec()
-        };
-
-        // encrypt payload and headers, and build the full message
-        let (intro_key, KeyContext { k_data, k_header_2 }) =
-            self.key_context.take().expect("to exist");
-
-        ChaChaPoly::with_nonce(k_data, pkt_num as u64)
-            .encrypt_with_ad_new(&header, &mut payload)
-            .expect("to succeed");
-
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        payload[payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(header.chunks_mut(8usize))
-            .zip([intro_key, *k_header_2])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-
-        let mut out = BytesMut::with_capacity(header.len() + payload.len());
-        out.put_slice(&header);
-        out.put_slice(&payload);
-
-        out
-    }
-}
-
-#[derive(Default)]
-pub struct TokenRequestBuilder {
-    dst_id: Option<u64>,
-    intro_key: Option<[u8; 32]>,
-    src_id: Option<u64>,
-}
-
-impl TokenRequestBuilder {
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, dst_id: u64) -> Self {
-        self.dst_id = Some(dst_id);
-        self
-    }
-
-    /// Specify source connection ID.
-    pub fn with_src_id(mut self, src_id: u64) -> Self {
-        self.src_id = Some(src_id);
-        self
-    }
-
-    /// Specify remote router's intro key.
-    pub fn with_intro_key(mut self, intro_key: [u8; 32]) -> Self {
-        self.intro_key = Some(intro_key);
-        self
-    }
-
-    /// Build [`TokenRequestBuilder`] into a byte vector.
-    pub fn build<R: Runtime>(mut self) -> BytesMut {
-        let intro_key = self.intro_key.take().expect("to exist");
-        let mut rng = R::rng();
-        let padding = {
-            let padding_len = rng.next_u32() % MAX_PADDING as u32 + 8;
-            let mut padding = vec![0u8; padding_len as usize];
-            rng.fill_bytes(&mut padding);
-
-            padding
-        };
-
-        let (mut header, pkt_num) = {
-            let mut out = BytesMut::with_capacity(LONG_HEADER_LEN);
-            let pkt_num = rng.next_u32();
-
-            out.put_u64_le(self.dst_id.take().expect("to exist"));
-            out.put_u32(pkt_num);
-            out.put_u8(*MessageType::TokenRequest);
-            out.put_u8(2u8); // version
-            out.put_u8(2u8); // net id TODO: make configurable
-            out.put_u8(0u8); // flag
-            out.put_u64_le(self.src_id.take().expect("to exist"));
-            out.put_u64(0u64);
-
-            (out, pkt_num)
-        };
-
-        let mut payload = Vec::with_capacity(10 + padding.len() + POLY13055_MAC_LEN);
-        payload.extend_from_slice(
-            &Block::DateTime {
-                timestamp: R::time_since_epoch().as_secs() as u32,
-            }
-            .serialize(),
-        );
-        payload.extend_from_slice(&Block::Padding { padding }.serialize());
-
-        // must succeed since all the parameters are controlled by us
-        ChaChaPoly::with_nonce(&intro_key, pkt_num as u64)
-            .encrypt_with_ad_new(&header, &mut payload)
-            .expect("to succeed");
-
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        payload[payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(header.chunks_mut(8usize))
-            .zip([intro_key, intro_key])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-
-        // encrypt last 16 bytes of the header
-        ChaCha::with_iv(intro_key, [0u8; IV_SIZE]).encrypt_ref(&mut header[16..32]);
-
-        let mut out = BytesMut::with_capacity(LONG_HEADER_LEN + payload.len());
-        out.put_slice(&header);
-        out.put_slice(&payload);
-
-        out
-    }
-}
-
-/// Unserialized `SessionCreated` message.
-pub struct SessionRequest {
-    /// Serialized, unencrypted header.
-    header: BytesMut,
-
-    /// Serialized, unencrypted payload
-    payload: Vec<u8>,
-}
-
-impl SessionRequest {
-    /// Get reference to header.
-    pub fn header(&self) -> &[u8] {
-        &self.header[..32]
-    }
-
-    /// Get reference to payload.
-    pub fn payload(&self) -> &[u8] {
-        &self.payload
-    }
-
-    /// Encrypt payload.
-    pub fn encrypt_payload(&mut self, cipher_key: &[u8], nonce: u64, state: &[u8]) {
-        // must succeed since all the parameters are controlled by us
-        ChaChaPoly::with_nonce(cipher_key, nonce)
-            .encrypt_with_ad_new(state, &mut self.payload)
-            .expect("to succeed");
-    }
-
-    /// Encrypt header.
-    pub fn encrypt_header(&mut self, k_header_1: [u8; 32], k_header_2: [u8; 32]) {
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        self.payload[self.payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(self.header.chunks_mut(8usize))
-            .zip([k_header_1, k_header_2])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-
-        // encrypt last 16 bytes of the header and the public key
-        ChaCha::with_iv(k_header_2, [0u8; IV_SIZE]).encrypt_ref(&mut self.header[16..64]);
-    }
-
-    /// Serialize [`SessionRequest`] into a byte vector.
-    pub fn build(self) -> BytesMut {
-        let mut out = BytesMut::with_capacity(self.header.len() + self.payload.len());
-        out.put_slice(&self.header);
-        out.put_slice(&self.payload);
-
-        out
-    }
-}
-
-#[derive(Default)]
-pub struct SessionRequestBuilder {
-    /// Destination connection ID.
-    dst_id: Option<u64>,
-
-    /// Local ephemeral public key.
-    ephemeral_key: Option<EphemeralPublicKey>,
-
-    /// Source connection ID.
-    src_id: Option<u64>,
-
-    /// Token.
-    token: Option<u64>,
-}
-
-impl SessionRequestBuilder {
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, dst_id: u64) -> Self {
-        self.dst_id = Some(dst_id);
-        self
-    }
-
-    /// Specify source connection ID.
-    pub fn with_src_id(mut self, src_id: u64) -> Self {
-        self.src_id = Some(src_id);
-        self
-    }
-
-    /// Specify token.
-    pub fn with_token(mut self, token: u64) -> Self {
-        self.token = Some(token);
-        self
-    }
-
-    /// Specify local ephemeral public key.
-    pub fn with_ephemeral_key(mut self, ephemeral_key: EphemeralPublicKey) -> Self {
-        self.ephemeral_key = Some(ephemeral_key);
-        self
-    }
-
-    /// Build [`SessionRequestBuilder`] into [`SessionRequest`].
-    pub fn build<R: Runtime>(mut self) -> SessionRequest {
-        let mut rng = R::rng();
-        let padding = {
-            let padding_len = rng.next_u32() % MAX_PADDING as u32 + 1;
-            let mut padding = vec![0u8; padding_len as usize];
-            rng.fill_bytes(&mut padding);
-
-            padding
-        };
-        let header = {
-            let mut out = BytesMut::with_capacity(LONG_HEADER_LEN + PUBLIC_KEY_LEN);
-
-            out.put_u64_le(self.dst_id.take().expect("to exist"));
-            out.put_u32(rng.next_u32());
-            out.put_u8(*MessageType::SessionRequest);
-            out.put_u8(2u8); // version
-            out.put_u8(2u8); // net id TODO: make configurable
-            out.put_u8(0u8); // flag
-            out.put_u64_le(self.src_id.take().expect("to exist"));
-            out.put_u64_le(self.token.take().expect("to exist"));
-            out.put_slice(self.ephemeral_key.take().expect("to exist").as_ref());
-
-            out
-        };
-
-        let mut payload = Vec::with_capacity(10 + padding.len() + POLY13055_MAC_LEN);
-        payload.extend_from_slice(
-            &Block::DateTime {
-                timestamp: R::time_since_epoch().as_secs() as u32,
-            }
-            .serialize(),
-        );
-        payload.extend_from_slice(&Block::Padding { padding }.serialize());
-
-        SessionRequest { header, payload }
-    }
-}
-
-/// Unserialized `SessionConfirmed` message.
-pub struct SessionConfirmed {
-    /// Serialized, unencrypted header.
-    header: BytesMut,
-
-    /// Serialized, unencrypted static key.
-    static_key: Vec<u8>,
-
-    /// Serialized, unecrypted payload.
-    payload: Vec<u8>,
-}
-
-impl SessionConfirmed {
-    /// Get reference to header.
-    pub fn header(&self) -> &[u8] {
-        &self.header[..16]
-    }
-
-    /// Get reference to public key.
-    pub fn public_key(&self) -> &[u8] {
-        &self.static_key
-    }
-
-    /// Encrypt public key.
-    pub fn encrypt_public_key(&mut self, cipher_key: &[u8], nonce: u64, state: &[u8]) {
-        // must succeed as the parameters are controlled by us
-        ChaChaPoly::with_nonce(cipher_key, nonce)
-            .encrypt_with_ad_new(state, &mut self.static_key)
-            .expect("to succeed");
-    }
-
-    /// Encrypt payload.
-    pub fn encrypt_payload(&mut self, cipher_key: &[u8], nonce: u64, state: &[u8]) {
-        // must succeed as the parameters are controlled by us
-        ChaChaPoly::with_nonce(cipher_key, nonce)
-            .encrypt_with_ad_new(state, &mut self.payload)
-            .expect("to succeed");
-    }
-
-    /// Encrypt header.
-    pub fn encrypt_header(&mut self, k_header_1: [u8; 32], k_header_2: [u8; 32]) {
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        self.payload[self.payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(self.header.chunks_mut(8usize))
-            .zip([k_header_1, k_header_2])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-    }
-
-    /// Serialize [`SessionConfirmed`] into a byte vector.
-    pub fn build(self) -> BytesMut {
-        let mut out =
-            BytesMut::with_capacity(self.header.len() + self.static_key.len() + self.payload.len());
-        out.put_slice(&self.header);
-        out.put_slice(&self.static_key);
-        out.put_slice(&self.payload);
-
-        out
-    }
-}
-
-/// `SessionConfirmed` builder.
-#[derive(Default)]
-pub struct SessionConfirmedBuilder {
-    /// Destination connection ID.
-    dst_id: Option<u64>,
-
-    /// Source connection ID.
-    src_id: Option<u64>,
-
-    /// Serialized local router info.
-    router_info: Option<Bytes>,
-
-    /// Local static public key.
-    static_key: Option<StaticPublicKey>,
-}
-
-impl SessionConfirmedBuilder {
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, dst_id: u64) -> Self {
-        self.dst_id = Some(dst_id);
-        self
-    }
-
-    /// Specify source connection ID.
-    pub fn with_src_id(mut self, src_id: u64) -> Self {
-        self.src_id = Some(src_id);
-        self
-    }
-
-    /// Specify router info.
-    pub fn with_router_info(mut self, router_info: Bytes) -> Self {
-        self.router_info = Some(router_info);
-        self
-    }
-
-    /// Specify local static public key.
-    pub fn with_static_key(mut self, static_key: StaticPublicKey) -> Self {
-        self.static_key = Some(static_key);
-        self
-    }
-
-    /// Build [`SessionConfirmedBuilder`] into a byte vector.
-    pub fn build(mut self) -> SessionConfirmed {
-        let header = {
-            let mut out = BytesMut::with_capacity(SHORT_HEADER_LEN);
-
-            out.put_u64_le(self.dst_id.take().expect("to exist"));
-            out.put_u32(0u32);
-            out.put_u8(*MessageType::SessionConfirmed);
-            out.put_u8(1u8); // 1 fragment
-            out.put_u16(0u16); // flags
-
-            out
-        };
-        let static_key = self.static_key.expect("to exist").to_vec();
-        let payload = {
-            let router_info = self.router_info.take().expect("to exist");
-            let mut out = BytesMut::with_capacity(5 + router_info.len());
-
-            out.put_u8(BlockType::RouterInfo.as_u8());
-            out.put_u16((2 + router_info.len()) as u16);
-            out.put_u8(0u8);
-            out.put_u8(1u8);
-            out.put_slice(&router_info);
-
-            out.to_vec()
-        };
-
-        SessionConfirmed {
-            header,
-            static_key,
-            payload,
-        }
-    }
-}
-
 /// Header kind.
 pub enum HeaderKind {
     /// Retry
@@ -1725,7 +1055,7 @@ pub struct HeaderReader<'a> {
 impl<'a> HeaderReader<'a> {
     /// Create new [`HeaderReader`].
     ///
-    /// Minimum size for `pkt` is 24 bytes as the IVs used for header decryption as 12 bytes long.
+    /// Minimum size for `pkt` is 24 bytes as the IVs used for header decryption are 12 bytes long.
     pub fn new(k_header_1: [u8; 32], pkt: &'a mut [u8]) -> Result<Self, Ssu2Error> {
         if pkt.len() < PKT_MIN_SIZE {
             return Err(Ssu2Error::NotEnoughBytes);
@@ -1900,288 +1230,5 @@ impl<'a> HeaderReader<'a> {
                 Err(Ssu2Error::UnexpectedMessage)
             }
         }
-    }
-}
-
-#[derive(Default)]
-pub struct RetryBuilder {
-    /// Remote's socket address.
-    address: Option<SocketAddr>,
-
-    /// Destination connection ID.
-    dst_id: Option<u64>,
-
-    /// Remote's intro key.
-    k_header_1: Option<[u8; 32]>,
-
-    /// Source connection ID.
-    src_id: Option<u64>,
-
-    /// Token.
-    token: Option<u64>,
-}
-
-impl RetryBuilder {
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, dst_id: u64) -> Self {
-        self.dst_id = Some(dst_id);
-        self
-    }
-
-    /// Specify source connection ID.
-    pub fn with_src_id(mut self, src_id: u64) -> Self {
-        self.src_id = Some(src_id);
-        self
-    }
-
-    /// Specify remote router's intro key.
-    pub fn with_k_header_1(mut self, k_header_1: [u8; 32]) -> Self {
-        self.k_header_1 = Some(k_header_1);
-        self
-    }
-
-    /// Specify token.
-    pub fn with_token(mut self, token: u64) -> Self {
-        self.token = Some(token);
-        self
-    }
-
-    /// Specify remote socket address.
-    pub fn with_address(mut self, address: SocketAddr) -> Self {
-        self.address = Some(address);
-        self
-    }
-
-    /// Build [`SessionConfirmedBuilder`] into a byte vector.
-    pub fn build<R: Runtime>(self) -> BytesMut {
-        let (mut header, pkt_num) = {
-            let mut out = BytesMut::with_capacity(LONG_HEADER_LEN);
-            let pkt_num = R::rng().next_u32();
-
-            out.put_u64_le(self.dst_id.expect("to exist"));
-            out.put_u32(pkt_num);
-            out.put_u8(*MessageType::Retry);
-            out.put_u8(2u8);
-            out.put_u8(2u8); // TODO: make configurable
-            out.put_u8(0u8);
-            out.put_u64_le(self.src_id.expect("to exist"));
-            out.put_u64(self.token.expect("to exist"));
-
-            (out, pkt_num)
-        };
-        let padding = {
-            let padding_len = R::rng().next_u32() as usize % MAX_PADDING + 1;
-            let mut padding = vec![0u8; padding_len];
-            R::rng().fill_bytes(&mut padding);
-
-            padding
-        };
-        let payload_size = 3 * 3 + 4 + 6 + padding.len() + POLY13055_MAC_LEN;
-        let k_header_1 = self.k_header_1.expect("to exist");
-
-        let mut payload = [
-            Block::DateTime {
-                timestamp: R::time_since_epoch().as_secs() as u32,
-            },
-            Block::Address {
-                address: self.address.expect("to exist"),
-            },
-            Block::Padding { padding },
-        ]
-        .into_iter()
-        .fold(BytesMut::with_capacity(payload_size), |mut out, block| {
-            out.put_slice(&block.serialize());
-            out
-        })
-        .to_vec();
-
-        // expected to succeed since the parameters are controlled by us
-        ChaChaPoly::with_nonce(&k_header_1, pkt_num as u64)
-            .encrypt_with_ad_new(&header, &mut payload)
-            .expect("to succeed");
-
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        payload[payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(header.chunks_mut(8usize))
-            .zip([k_header_1, k_header_1])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-
-        // encrypt third part of the header
-        ChaCha::with_iv(k_header_1, [0u8; IV_SIZE]).encrypt_ref(&mut header[16..32]);
-
-        let mut out = BytesMut::with_capacity(header.len() + payload.len());
-        out.put_slice(&header);
-        out.put_slice(&payload);
-
-        out
-    }
-}
-
-/// Unserialized `SessionCreated` message.
-pub struct SessionCreated {
-    /// Serialized, unencrypted header.
-    header: BytesMut,
-
-    /// Serialized, unencrypted payload
-    payload: Vec<u8>,
-}
-
-impl SessionCreated {
-    /// Get reference to header.
-    pub fn header(&self) -> &[u8] {
-        &self.header[..32]
-    }
-
-    /// Get reference to payload.
-    pub fn payload(&self) -> &[u8] {
-        &self.payload
-    }
-
-    /// Encrypt header.
-    pub fn encrypt_header(&mut self, k_header_1: [u8; 32], k_header_2: [u8; 32]) {
-        // encrypt first 16 bytes of the long header
-        //
-        // https://geti2p.net/spec/ssu2#header-encryption-kdf
-        self.payload[self.payload.len() - 2 * IV_SIZE..]
-            .chunks(IV_SIZE)
-            .zip(self.header.chunks_mut(8usize))
-            .zip([k_header_1, k_header_2])
-            .for_each(|((chunk, header_chunk), key)| {
-                ChaCha::with_iv(
-                    key,
-                    TryInto::<[u8; IV_SIZE]>::try_into(chunk).expect("to succeed"),
-                )
-                .decrypt([0u8; 8])
-                .iter()
-                .zip(header_chunk.iter_mut())
-                .for_each(|(mask_byte, header_byte)| {
-                    *header_byte ^= mask_byte;
-                });
-            });
-
-        ChaCha::with_iv(k_header_2, [0u8; IV_SIZE]).encrypt_ref(&mut self.header[16..64]);
-    }
-
-    /// Encrypt payload.
-    pub fn encrypt_payload(&mut self, cipher_key: &[u8], nonce: u64, state: &[u8]) {
-        // expected to succeed as the parameters are controlled by us
-        ChaChaPoly::with_nonce(cipher_key, nonce)
-            .encrypt_with_ad_new(state, &mut self.payload)
-            .expect("to succeed");
-    }
-
-    /// Serialize [`SessionCreated`] into a byte vector.
-    pub fn build(self) -> BytesMut {
-        let mut out = BytesMut::with_capacity(self.header.len() + self.payload.len());
-        out.put_slice(&self.header);
-        out.put_slice(&self.payload);
-
-        out
-    }
-}
-
-/// `SessionCreated` message builder.
-#[derive(Default)]
-pub struct SessionCreatedBuilder {
-    /// Remote router's address.
-    address: Option<SocketAddr>,
-
-    /// Destination connection ID.
-    dst_id: Option<u64>,
-
-    /// Our ephemeral public key.
-    ephemeral_key: Option<EphemeralPublicKey>,
-
-    /// Source connection ID.
-    src_id: Option<u64>,
-}
-
-impl SessionCreatedBuilder {
-    /// Specify remote's socket address.
-    pub fn with_address(mut self, address: SocketAddr) -> Self {
-        self.address = Some(address);
-        self
-    }
-
-    /// Specify destination connection ID.
-    pub fn with_dst_id(mut self, dst_id: u64) -> Self {
-        self.dst_id = Some(dst_id);
-        self
-    }
-
-    /// Specify local ephemeral public key.
-    pub fn with_ephemeral_key(mut self, ephemeral_key: EphemeralPublicKey) -> Self {
-        self.ephemeral_key = Some(ephemeral_key);
-        self
-    }
-
-    /// Specify source connection ID.
-    pub fn with_src_id(mut self, src_id: u64) -> Self {
-        self.src_id = Some(src_id);
-        self
-    }
-
-    /// Build [`SessionCreatedBuilder`] into [`SessionCreated`] by creating a long header
-    /// and a payload with needed blocks.
-    ///
-    /// This function doesn't return a serialized `SessionCreated` message as the caller needs to
-    /// encrypt the payload with "non-static" key/state which future encryption/decryption is
-    /// depended on.
-    pub fn build<R: Runtime>(mut self) -> SessionCreated {
-        let header = {
-            let mut out = BytesMut::with_capacity(LONG_HEADER_LEN);
-
-            out.put_u64_le(self.dst_id.expect("to exist"));
-            out.put_u32(R::rng().next_u32());
-            out.put_u8(*MessageType::SessionCreated);
-            out.put_u8(2u8);
-            out.put_u8(2u8); // TODO: make configurable
-            out.put_u8(0u8);
-            out.put_u64_le(self.src_id.expect("to exist"));
-            out.put_u64(0u64);
-            out.put_slice(self.ephemeral_key.take().expect("to exist").as_ref());
-
-            out
-        };
-        let padding = {
-            let padding_len = R::rng().next_u32() as usize % MAX_PADDING + 1;
-            let mut padding = vec![0u8; padding_len];
-            R::rng().fill_bytes(&mut padding);
-
-            padding
-        };
-        let payload_size = 3 * 3 + 4 + 6 + padding.len() + POLY13055_MAC_LEN;
-
-        let payload = [
-            Block::DateTime {
-                timestamp: R::time_since_epoch().as_secs() as u32,
-            },
-            Block::Address {
-                address: self.address.expect("to exist"),
-            },
-            Block::Padding { padding },
-        ]
-        .into_iter()
-        .fold(BytesMut::with_capacity(payload_size), |mut out, block| {
-            out.put_slice(&block.serialize());
-            out
-        })
-        .to_vec();
-
-        SessionCreated { header, payload }
     }
 }
